+++ apps/web/src/shared/ui/chat/ChatListContent.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { useChatList } from "@/shared/api/chat/use-chat-list";
-import type { ChatEmptyAction } from "./ChatRoomListPanel";
 import { ChatRoomListPanel } from "./ChatRoomListPanel";
+import type { ChatEmptyAction } from "./ChatRoomsEmpty";
 
 export interface ChatListContentProps {
   /** 방 링크 베이스(예 "/customer/chat") — 역할별 경로 */
